#18 refactor: 세금계산서 승인 상태 추가
