--- src/main/java/com/seoulmilk/seoulmilkServer/domain/ntsTax/domain/ARAP.java
@@ -1,5 +1,0 @@
-package com.seoulmilk.seoulmilkServer.domain.ntsTax.domain;
-
-public enum ARAP {
-    AR, AP
-}
