+++ src/main/java/com/seoulmilk/seoulmilkServer/domain/ntsTax/domain/enums/Status.java
@@ -1,0 +1,5 @@
+package com.seoulmilk.seoulmilkServer.domain.ntsTax.domain.enums;
+
+public enum Status {
+    APPROVAL, REJECTION, WAITING
+}
