+++ src/main/java/com/seoulmilk/seoulmilkServer/domain/ntsTax/dto/response/UpdateNtsTaxResponseDTO.java
@@ -1,6 +1,6 @@
 package com.seoulmilk.seoulmilkServer.domain.ntsTax.dto.response;
 
-import com.seoulmilk.seoulmilkServer.domain.ntsTax.domain.ARAP;
+import com.seoulmilk.seoulmilkServer.domain.ntsTax.domain.enums.ARAP;
 import com.seoulmilk.seoulmilkServer.domain.ntsTax.domain.NtsTax;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
